+++ sodium/widgets/__init__.py
@@ -4,5 +4,6 @@ Widgets module
 from .button import *
 from .freelabel  import *
 from .label import *
+from .frame import *
 
-__all__ = ["Button", "FreeLabel", "Label"]
+__all__ = ["Button", "FreeLabel", "Label", "Frame"]